remove Enum extensions, src/prelude, rename OffcanvasPlacement -> Placement
