+++ lib/bootstrap/icons.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: constant_identifier_names
 
-import 'bootstrap_core.dart';
+import '../src/prelude.dart';
 
 DeactNode icon(
   BIcon icon, {
